backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sgfmill>=1.1.1
litellm>=1.30.0
pyyaml>=6.0
python-multipart>=0.0.6
aiofiles>=23.1.0
